scraper/requirements.txt
PyGithub~=1.51
python-dotenv~=0.13.0
backports.zoneinfo~=0.2.1
requests~=2.24.0
schedule~=0.6.0
